Lift shell-model values + UI policy types into core packages (iOS refactor 5079 wave 1)
Extract the pure MobileShellModels value types + MobileShellRouteAuthPolicy
into a new Foundation-only CmuxMobileShellModel core package, and the pure
presentation/layout policy types (navigation, layout, safe-area, root auth
gate, sign-in code) into a new CmuxMobileWorkspace package, one type per file
with DocC. cmuxFeature now depends on both; MobileRootAuthGate.syncShell stays
in the feature target as an extension since it reaches the god store. Moved the
matching tests into per-package @Suites.
