+++ Packages/CmuxMobileShellModel/Package.swift
@@ -1,0 +1,42 @@
+// swift-tools-version: 6.0
+
+import PackageDescription
+
+let package = Package(
+    name: "CmuxMobileShellModel",
+    platforms: [
+        .iOS(.v18),
+        .macOS(.v14),
+    ],
+    products: [
+        .library(
+            name: "CmuxMobileShellModel",
+            targets: ["CmuxMobileShellModel"]
+        ),
+    ],
+    dependencies: [
+        .package(path: "../CMUXMobileCore"),
+    ],
+    targets: [
+        .target(
+            name: "CmuxMobileShellModel",
+            dependencies: [
+                "CMUXMobileCore",
+            ],
+            swiftSettings: [
+                .swiftLanguageMode(.v6),
+                .enableUpcomingFeature("ExistentialAny"),
+                .enableUpcomingFeature("InternalImportsByDefault"),
+            ]
+        ),
+        .testTarget(
+            name: "CmuxMobileShellModelTests",
+            dependencies: ["CmuxMobileShellModel"],
+            swiftSettings: [
+                .swiftLanguageMode(.v6),
+                .enableUpcomingFeature("ExistentialAny"),
+                .enableUpcomingFeature("InternalImportsByDefault"),
+            ]
+        ),
+    ]
+)
