+++ Packages/CmuxMobileShellModel/Sources/CmuxMobileShellModel/MobileConnectionState.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+/// The connection state of the mobile shell to a remote Mac.
+public enum MobileConnectionState: Equatable, Sendable {
+    /// No active connection.
+    case disconnected
+    /// An active connection to a remote Mac.
+    case connected
+}
