+++ Packages/CmuxMobileShellModel/Sources/CmuxMobileShellModel/MobileWorkspacePreview.swift
@@ -1,0 +1,45 @@
+import Foundation
+
+/// A lightweight, `Sendable` snapshot of a remote workspace shown in the mobile shell.
+///
+/// This is a pure value model: it carries the workspace identity, display name, and
+/// the ordered list of its terminals. It is decoupled from any connection, RPC, or
+/// rendering concern so that both the domain coordinators and the SwiftUI layer can
+/// consume the same immutable shape.
+public struct MobileWorkspacePreview: Identifiable, Equatable, Sendable {
+    /// A stable, string-backed identifier for a ``MobileWorkspacePreview``.
+    public struct ID: RawRepresentable, Hashable, Codable, Sendable, ExpressibleByStringLiteral {
+        /// The underlying workspace identifier string.
+        public var rawValue: String
+
+        /// Creates an identifier from its raw string value.
+        /// - Parameter rawValue: The backing workspace identifier.
+        public init(rawValue: String) {
+            self.rawValue = rawValue
+        }
+
+        /// Creates an identifier from a string literal.
+        /// - Parameter value: The backing workspace identifier.
+        public init(stringLiteral value: String) {
+            self.rawValue = value
+        }
+    }
+
+    /// The workspace's stable identifier.
+    public var id: ID
+    /// The workspace's user-facing display name.
+    public var name: String
+    /// The terminals contained in the workspace, in display order.
+    public var terminals: [MobileTerminalPreview]
+
+    /// Creates a workspace preview.
+    /// - Parameters:
+    ///   - id: The workspace's stable identifier.
+    ///   - name: The workspace's user-facing display name.
+    ///   - terminals: The terminals contained in the workspace, in display order.
+    public init(id: ID, name: String, terminals: [MobileTerminalPreview]) {
+        self.id = id
+        self.name = name
+        self.terminals = terminals
+    }
+}
