+++ ios/cmuxPackage/Package.swift
@@ -19,9 +19,11 @@ let package = Package(
         .package(path: "../../Packages/CMUXMobileCore"),
         .package(path: "../../Packages/CmuxMobileAuth"),
         .package(path: "../../Packages/CmuxMobileDiagnostics"),
+        .package(path: "../../Packages/CmuxMobileShellModel"),
         .package(path: "../../Packages/CmuxMobileSupport"),
         .package(path: "../../Packages/CmuxMobileTerminal"),
         .package(path: "../../Packages/CmuxMobileTransport"),
+        .package(path: "../../Packages/CmuxMobileWorkspace"),
         .package(path: "../../vendor/stack-auth-swift-sdk-prerelease"),
     ],
     targets: [
@@ -32,9 +34,11 @@ let package = Package(
                 "CMUXMobileCore",
                 "CmuxMobileAuth",
                 "CmuxMobileDiagnostics",
+                "CmuxMobileShellModel",
                 "CmuxMobileSupport",
                 "CmuxMobileTerminal",
                 "CmuxMobileTransport",
+                "CmuxMobileWorkspace",
                 .product(name: "StackAuth", package: "stack-auth-swift-sdk-prerelease"),
             ],
             swiftSettings: [
@@ -50,9 +54,11 @@ let package = Package(
                 "CMUXMobileCore",
                 "CmuxMobileAuth",
                 "CmuxMobileDiagnostics",
+                "CmuxMobileShellModel",
                 "CmuxMobileSupport",
                 "CmuxMobileTerminal",
                 "CmuxMobileTransport",
+                "CmuxMobileWorkspace",
                 .product(name: "StackAuth", package: "stack-auth-swift-sdk-prerelease"),
             ],
             swiftSettings: [
