+++ ios/cmuxPackage/Sources/cmuxFeature/PairingView.swift
@@ -2,6 +2,7 @@ import Foundation
 @preconcurrency import AVFoundation
 import CMUXMobileCore
 import CmuxMobileAuth
+import CmuxMobileShellModel
 import CmuxMobileSupport
 import CmuxMobileTerminal
 import Observation
